make method call accept options
The original ordering of the arguments was, IMO, incorrect. It was very
difficult to pass a timestamp, as that was the final argument. (I had a
requirement to anchor a batch of values to the same timestamp.)

The new approach expects two mandatory arguments, for metric value and name, and
lets the user override host, tags, and timestamp via an options hash. This is a
breaking change for anyone using the writer API prior to this version.